typecasts for prev commit (splint)

+++ src/cli/modes/mode_decode.c
@@ -142,7 +142,7 @@ mode_decode(const uint optind, const uint argc, char *const *const argv)
 	// exit if any errors
 	if UNLIKELY ( nerrors_file != 0 ){
 		if ( nerrors_file > (size_t) 255u ){
-			nerrors_file = 255u;
+			nerrors_file = (size_t) 255u;
 		}
 		exit((int) nerrors_file);
 	}
